refactor mp weights tests filename
